backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
yfinance==0.2.38
pandas==2.2.1
numpy==1.26.4
ta==0.11.0
feedparser==6.0.11
pyjwt==2.8.0
python-multipart==0.0.9
pydantic[email]==2.6.4
httpx==0.27.0
twilio==8.13.0
python-dotenv==1.0.1
